requirements.txt
dash
dash[testing]
pandas
python-magic